Add time and IP handlers in internal/api module
This commit creates two new files in the internal/api module: time.go and ip.go. Time.go adds a method to get the current time from a server and an endpoint to fetch this time. Ip.go introduces handlers for getting the requester's IP in different formats (both IPv4 and IPv6).